requirements.txt
machinevision-toolbox-python==0.5.5
numpy==1.23.1
requests==2.31.0
pygame==2.1.2
torch==1.12.1
torchvision==0.13.1
matplotlib==3.5.2
tqdm==4.64.0
h5py==3.7.0
